requirements.txt
requests
flask_cors
flask
zappa
mysql-connector-python
dotenv